requirements.txt
dash
numpy
pandas
SQLAlchemy
mysql-connector-python
pymysql
dash-bootstrap-components
